backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

supabase==2.0.0
gotrue==1.3.1
postgrest==0.13.0
storage3==0.5.3
realtime==1.0.6

pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7

langchain==0.1.17
langchain-core==0.1.53
langchain-community==0.0.38
langchain-pinecone==0.1.3
langchain-google-genai==0.0.6
langchain-text-splitters==0.0.2
langsmith==0.1.147

sentence-transformers==5.1.2
transformers==4.57.3
torch==2.9.1
tokenizers==0.22.1
numpy==1.26.4
scipy==1.16.3
scikit-learn==1.4.2

google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
google-auth==2.43.0
google-api-core==2.28.1
googleapis-common-protos==1.72.0

httpx==0.24.1
requests==2.31.0
urllib3==2.5.0
idna==3.11
charset-normalizer==3.4.4

pillow==12.0.0

pydantic==2.5.0
orjson==3.11.4
typing_extensions==4.15.0

python-jose==3.3.0
cryptography==46.0.3
rsa==4.9.1